Make mure refresh complete in one go
As described in https://github.com/kitsuyui/mure/issues/85 , the current `mure refresh` does not fully update the repository unless executed twice.
Let's solve this.

Specifically, postpone the work of updating the default branch.
Also, return PullFastForwardStatus so that it is possible to tell whether it was Fast-Forward merged or already in the latest state.