Fix error in repository without codecov enabled
When `mure issues` is executed, an error occurs in a repository without codecov enabled.
Fix it.
It is useless to get coverage for a repository that does not have codecov enabled, so ignore such repositories.
There was a problem with the implementation of https://crates.io/crates/codecov, so I upgraded it to a version that solved it.